requirements.txt
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
pyarrow>=14.0.0
statsmodels>=0.14.0
python-dateutil>=2.8.0
